+++ src/application/dtos/transaction.go
@@ -1,15 +1,15 @@
 package dtos
 
 type TransactionRequest struct {
-	AccountID       int64  `json:"account_id" binding:"required"`
-	OperationTypeID int8   `json:"operation_type_id" binding:"required"`
-	Amount          uint64 `json:"amount" binding:"-"`
+	AccountID       int64   `json:"account_id" binding:"required"`
+	OperationTypeID int8    `json:"operation_type_id" binding:"required"`
+	Amount          float64 `json:"amount" binding:"-"`
 }
 
 type TransactionResponse struct {
-	ID              int64  `json:"id"`
-	AccountID       int64  `json:"account_id"`
-	OperationTypeID int8   `json:"operation_type_id"`
-	Amount          uint64 `json:"amount"`
-	EventDate       string `json:"event_date"`
+	ID              int64   `json:"id"`
+	AccountID       int64   `json:"account_id"`
+	OperationTypeID int8    `json:"operation_type_id"`
+	Amount          float64 `json:"amount"`
+	EventDate       string  `json:"event_date"`
 }
